Promore StableConnectIdentities feature gate to beta and enable it by default
