+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaMirrorMaker2AssemblyOperatorMockTest.java
@@ -13,10 +13,9 @@ import io.strimzi.api.kafka.model.KafkaMirrorMaker2;
 import io.strimzi.api.kafka.model.KafkaMirrorMaker2Builder;
 import io.strimzi.api.kafka.model.KafkaMirrorMaker2Resources;
 import io.strimzi.api.kafka.model.status.Condition;
-import io.strimzi.platform.KubernetesVersion;
-import io.strimzi.operator.cluster.PlatformFeaturesAvailability;
 import io.strimzi.operator.cluster.ClusterOperatorConfig;
 import io.strimzi.operator.cluster.KafkaVersionTestUtils;
+import io.strimzi.operator.cluster.PlatformFeaturesAvailability;
 import io.strimzi.operator.cluster.ResourceUtils;
 import io.strimzi.operator.cluster.model.KafkaVersion;
 import io.strimzi.operator.cluster.operator.resource.DefaultZookeeperScalerProvider;
@@ -25,7 +24,9 @@ import io.strimzi.operator.cluster.operator.resource.ZookeeperLeaderFinder;
 import io.strimzi.operator.common.BackOff;
 import io.strimzi.operator.common.DefaultAdminClientProvider;
 import io.strimzi.operator.common.Reconciliation;
+import io.strimzi.operator.common.model.Labels;
 import io.strimzi.operator.common.model.OrderedProperties;
+import io.strimzi.platform.KubernetesVersion;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.mockkube2.MockKube2;
 import io.vertx.core.Future;
@@ -40,6 +41,7 @@ import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -52,7 +54,6 @@ import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -62,21 +63,22 @@ import static org.mockito.Mockito.when;
 @EnableKubernetesMockClient(crud = true)
 @ExtendWith(VertxExtension.class)
 public class KafkaMirrorMaker2AssemblyOperatorMockTest {
-
     private static final Logger LOGGER = LogManager.getLogger(KafkaMirrorMaker2AssemblyOperatorMockTest.class);
-
-    private static final KafkaVersion.Lookup VERSIONS = KafkaVersionTestUtils.getKafkaVersionLookup();
 
     private static final String NAMESPACE = "my-namespace";
     private static final String CLUSTER_NAME = "my-mm2-cluster";
-    private static WorkerExecutor sharedWorkerExecutor;
+    private static final int REPLICAS = 3;
+    private static final KafkaVersion.Lookup VERSIONS = KafkaVersionTestUtils.getKafkaVersionLookup();
+    private static final PlatformFeaturesAvailability PFA = new PlatformFeaturesAvailability(false, KubernetesVersion.MINIMAL_SUPPORTED_VERSION);
 
-    private final int replicas = 3;
+    private static WorkerExecutor sharedWorkerExecutor;
 
     // Injected by Fabric8 Mock Kubernetes Server
     @SuppressWarnings("unused")
     private KubernetesClient client;
     private MockKube2 mockKube;
+    private ResourceOperatorSupplier supplier;
+    private StrimziPodSetController podSetController;
 
     private static Vertx vertx;
     private KafkaMirrorMaker2AssemblyOperator kco;
@@ -93,36 +95,39 @@ public class KafkaMirrorMaker2AssemblyOperatorMockTest {
         vertx.close();
     }
 
-    private void setMirrorMaker2Resource(KafkaMirrorMaker2 mirrorMaker2Resource) {
+    @BeforeEach
+    public void beforeEach() {
         // Configure the Kubernetes Mock
         mockKube = new MockKube2.MockKube2Builder(client)
+                .withKafkaConnectCrd()
+                .withKafkaCrd()
                 .withKafkaMirrorMaker2Crd()
                 .withStrimziPodSetCrd()
-                .withInitialKafkaMirrorMaker2s(mirrorMaker2Resource)
-                .withDeploymentController()
+                .withPodController()
                 .build();
         mockKube.start();
+
+        supplier = new ResourceOperatorSupplier(vertx, client,
+                new ZookeeperLeaderFinder(vertx,
+                        // Retry up to 3 times (4 attempts), with overall max delay of 35000ms
+                        () -> new BackOff(5_000, 2, 4)),
+                new DefaultAdminClientProvider(),
+                new DefaultZookeeperScalerProvider(),
+                ResourceUtils.metricsProvider(),
+                PFA, 60_000L);
+        podSetController = new StrimziPodSetController(NAMESPACE, Labels.EMPTY, supplier.kafkaOperator, supplier.connectOperator, supplier.mirrorMaker2Operator, supplier.strimziPodSetOperator, supplier.podOperations, supplier.metricsProvider, Integer.parseInt(ClusterOperatorConfig.POD_SET_CONTROLLER_WORK_QUEUE_SIZE.defaultValue()));
+        podSetController.start();
     }
 
     @AfterEach
     public void afterEach() {
+        podSetController.stop();
         mockKube.stop();
     }
 
-
     private Future<Void> createMirrorMaker2Cluster(VertxTestContext context, KafkaConnectApi kafkaConnectApi, boolean reconciliationPaused) {
-        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.MINIMAL_SUPPORTED_VERSION);
-        ResourceOperatorSupplier supplier = new ResourceOperatorSupplier(vertx, client,
-                new ZookeeperLeaderFinder(vertx,
-                    // Retry up to 3 times (4 attempts), with overall max delay of 35000ms
-                    () -> new BackOff(5_000, 2, 4)),
-                new DefaultAdminClientProvider(),
-                new DefaultZookeeperScalerProvider(),
-                ResourceUtils.metricsProvider(),
-                pfa, 60_000L);
-
         ClusterOperatorConfig config = ResourceUtils.dummyClusterOperatorConfig(VERSIONS);
-        kco = new KafkaMirrorMaker2AssemblyOperator(vertx, pfa,
+        kco = new KafkaMirrorMaker2AssemblyOperator(vertx, PFA,
             supplier,
             config,
             foo -> kafkaConnectApi);
@@ -133,31 +138,34 @@ public class KafkaMirrorMaker2AssemblyOperatorMockTest {
         kco.reconcile(new Reconciliation("test-trigger", KafkaMirrorMaker2.RESOURCE_KIND, NAMESPACE, CLUSTER_NAME))
             .onComplete(context.succeeding(ar -> context.verify(() -> {
                 if (!reconciliationPaused) {
-                    assertThat(client.apps().deployments().inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.deploymentName(CLUSTER_NAME)).get(), is(notNullValue()));
+                    assertThat(Crds.strimziPodSetOperation(client).inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.deploymentName(CLUSTER_NAME)).get(), is(notNullValue()));
+                    assertThat(client.apps().deployments().inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.deploymentName(CLUSTER_NAME)).get(), is(nullValue()));
                     assertThat(client.configMaps().inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.metricsAndLogConfigMapName(CLUSTER_NAME)).get(), is(notNullValue()));
                     assertThat(client.services().inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.serviceName(CLUSTER_NAME)).get(), is(notNullValue()));
                     assertThat(client.policy().v1().podDisruptionBudget().inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.deploymentName(CLUSTER_NAME)).get(), is(notNullValue()));
                 } else {
-                    assertThat(client.apps().deployments().inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.deploymentName(CLUSTER_NAME)).get(), is(nullValue()));
+                    assertThat(Crds.strimziPodSetOperation(client).inNamespace(NAMESPACE).withName(KafkaMirrorMaker2Resources.deploymentName(CLUSTER_NAME)).get(), is(nullValue()));
                 }
 
                 created.complete();
             })));
+
         return created.future();
     }
 
     @Test
     public void testReconcileUpdate(VertxTestContext context) {
-        setMirrorMaker2Resource(new KafkaMirrorMaker2Builder()
+        Crds.kafkaMirrorMaker2Operation(client).resource(new KafkaMirrorMaker2Builder()
                 .withMetadata(new ObjectMetaBuilder()
                         .withName(CLUSTER_NAME)
                         .withNamespace(NAMESPACE)
                         .withLabels(TestUtils.map("foo", "bar"))
                         .build())
                 .withNewSpec()
-                .withReplicas(replicas)
+                .withReplicas(REPLICAS)
                 .endSpec()
-            .build());
+            .build()).create();
+
         KafkaConnectApi mock = mock(KafkaConnectApi.class);
         when(mock.list(any(), anyString(), anyInt())).thenReturn(Future.succeededFuture(emptyList()));
         when(mock.updateConnectLoggers(any(), anyString(), anyInt(), anyString(), any(OrderedProperties.class))).thenReturn(Future.succeededFuture());
@@ -174,7 +182,7 @@ public class KafkaMirrorMaker2AssemblyOperatorMockTest {
 
     @Test
     public void testPauseReconcile(VertxTestContext context) {
-        setMirrorMaker2Resource(new KafkaMirrorMaker2Builder()
+        Crds.kafkaMirrorMaker2Operation(client).resource(new KafkaMirrorMaker2Builder()
                 .withMetadata(new ObjectMetaBuilder()
                         .withName(CLUSTER_NAME)
                         .withNamespace(NAMESPACE)
@@ -182,9 +190,10 @@ public class KafkaMirrorMaker2AssemblyOperatorMockTest {
                         .withAnnotations(singletonMap("strimzi.io/pause-reconciliation", "true"))
                         .build())
                 .withNewSpec()
-                .withReplicas(replicas)
+                .withReplicas(REPLICAS)
                 .endSpec()
-                .build());
+                .build()).create();
+
         KafkaConnectApi mock = mock(KafkaConnectApi.class);
         when(mock.list(any(), anyString(), anyInt())).thenReturn(Future.succeededFuture(emptyList()));
         when(mock.updateConnectLoggers(any(), anyString(), anyInt(), anyString(), any(OrderedProperties.class))).thenReturn(Future.succeededFuture());
@@ -199,7 +208,7 @@ public class KafkaMirrorMaker2AssemblyOperatorMockTest {
                     Resource<KafkaMirrorMaker2> resource = Crds.kafkaMirrorMaker2Operation(client).inNamespace(NAMESPACE).withName(CLUSTER_NAME);
 
                     if (resource.get().getStatus() == null) {
-                        fail();
+                        context.failNow("Status is null");
                     }
                     List<Condition> conditions = resource.get().getStatus().getConditions();
                     boolean conditionFound = false;
