+++ documentation/modules/operators/ref-operator-cluster-feature-gate-releases.adoc
@@ -21,8 +21,8 @@ Alpha and beta stage features are removed if they do not prove to be useful.
 * The `ServiceAccountPatching` feature gate moved to GA stage in Strimzi 0.30. It is now permanently enabled and cannot be disabled.
 * The `UseStrimziPodSets` feature gate moved to GA stage in Strimzi 0.35 and the support for StatefulSets is completely removed. It is now permanently enabled and cannot be disabled.
 * The `UseKRaft` feature gate is available for development only and does not currently have a planned release for moving to the beta phase.
-* The `StableConnectIdentities` feature gate is in alpha stage and is disabled by default.
-  It is expected to move to beta phase and be enabled by default from Strimzi 0.37.
+* The `StableConnectIdentities` feature gate is in beta stage and is enabled by default.
+  It is expected to move to GA phase and be always enabled from Strimzi 0.39.
 * The `KafkaNodePools` feature gate is in alpha stage and is disabled by default.
   It is expected to move to beta phase and be enabled by default from Strimzi 0.39.
 
@@ -59,8 +59,8 @@ NOTE: Feature gates might be removed when they reach GA. This means that the fea
 
 ¦`StableConnectIdentities`
 ¦0.34
-¦0.37 (planned)
-¦ -
+¦0.37
+¦0.39 (planned)
 
 ¦`KafkaNodePools`
 ¦0.36
