+++ systemtest/src/test/resources/upgrade/BundleUpgrade.yaml
@@ -62,5 +62,5 @@
     status: stable
     flakyEnvVariable: none
     reason: Test is working on environment, where k8s server version is < 1.22
-  featureGatesBefore: ""
+  featureGatesBefore: "-StableConnectIdentities"
   featureGatesAfter: "+StableConnectIdentities"
